+++ packages/coding-standards/index.mjs
@@ -1,4 +1,0 @@
-import { default as cspellConfig } from "./cspell/index.mjs";
-import { default as prettierConfig } from "./prettier/index.mjs";
-import { default as stylelintConfig } from "./stylelint/index.mjs";
-export { cspellConfig, prettierConfig, stylelintConfig };